requirements.txt
# Core FastAPI dependencies
fastapi==0.115.12
uvicorn[standard]==0.34.3
pydantic==2.11.5
pydantic-settings==2.9.1

# HTTP Client for microservices communication
httpx==0.28.1
aiohttp==3.12.9

# Service Discovery & Health Checks
python-consul==1.1.0

# Message Queue
aio-pika==9.5.5
aiokafka==0.12.0

# Caching
redis==6.2.0

# Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Database
motor==3.6.0
pymongo==4.9.1

# Templates and Static Files
jinja2==3.1.4
aiofiles==24.1.0

# Utilities
python-dotenv==1.1.0

# Monitoring and Logging
prometheus-client==0.22.1
structlog==24.4.0
psutil==7.0.0

# Rate Limiting
slowapi==0.1.9

# Development
pytest==8.4.0
pytest-asyncio==0.26.0